+++ EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -73,17 +73,7 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
         store.completeRetrieval(with: anyData())
         XCTAssertTrue(receivedMsg.isEmpty)
     }
-    
-    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
-        let (sut, store) = makeSUT()
-        let url = anyURL()
-        let data = anyData()
-        sut.save(data, for: url) { _ in
-            
-        }
-        XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
-    }
-    
+        
     //MARK: - Helpers
     private func makeSUT(currentDate: @escaping () -> Date = Date.init, file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
         let store = FeedImageDataStoreSpy()
